Contact and Send Message Fragment + Investment Chart

+++ EasyinvestTest/app/src/main/java/br/heitor/easyinvest/model/Graph.java
@@ -3,23 +3,23 @@ package br.heitor.easyinvest.model;
 import java.util.ArrayList;
 
 public class Graph {
-    private ArrayList<Double> CDI;
-    private ArrayList<Double> fund;
+    private ArrayList<Float> CDI;
+    private ArrayList<Float> fund;
     private ArrayList<String> x;
 
-    public ArrayList<Double> getCDI() {
+    public ArrayList<Float> getCDI() {
         return this.CDI;
     }
 
-    public void setCDI(ArrayList<Double> CDI) {
+    public void setCDI(ArrayList<Float> CDI) {
         this.CDI = CDI;
     }
 
-    public ArrayList<Double> getFund() {
+    public ArrayList<Float> getFund() {
         return this.fund;
     }
 
-    public void setFund(ArrayList<Double> fund) {
+    public void setFund(ArrayList<Float> fund) {
         this.fund = fund;
     }
